+++ src/Entities/Customer.php
@@ -11,6 +11,7 @@
 
 namespace Rebilly\Entities;
 
+use Rebilly\Entities\Contact\Email;
 use Rebilly\Rest\Entity;
 
 /**
@@ -31,57 +32,125 @@ use Rebilly\Rest\Entity;
 final class Customer extends Entity
 {
     /**
-     * @return string
+     * @deprecated Use getPrimaryAddress()->getEmails() instead
+     *
+     * @return string|null
      */
     public function getEmail()
     {
-        return $this->getAttribute('email');
+        $primaryAddress = $this->getPrimaryAddress();
+        if ($primaryAddress === null) {
+            return null;
+        }
+
+        $emails = $primaryAddress->getEmails();
+        if (empty($emails)) {
+            return null;
+        }
+
+        foreach ($emails as $email) {
+            if (is_array($email) && $email['primary'] === true) {
+                return $email['value'];
+            }
+
+            return $email->getValue();
+        }
+
+        return $emails[0]->getValue();
     }
 
     /**
+     * @deprecated Use getPrimaryAddress()->setEmails() instead
+     *
      * @param string $value
      *
      * @return $this
      */
     public function setEmail($value)
     {
-        return $this->setAttribute('email', $value);
+        $primaryAddress = $this->getPrimaryAddress();
+        $email = Email::createFromData([
+            'label' => 'main',
+            'value' => $value,
+            'primary' => true,
+        ]);
+
+        if ($primaryAddress !== null) {
+            $primaryAddress->setEmails([$email]);
+        } else {
+            $primaryAddress = Address::createFromData(['emails' => [$email]]);
+        }
+
+        $this->setPrimaryAddress($primaryAddress);
+        $this->setAttribute('email', $value);
+
+        return $this;
     }
 
     /**
+     * @deprecated Use getPrimaryAddress()->getFirstName() instead
+     *
      * @return string
      */
     public function getFirstName()
     {
-        return $this->getAttribute('firstName');
+        return $this->getPrimaryAddress() !== null ? $this->getPrimaryAddress()->getFirstName() : null;
     }
 
     /**
+     * @deprecated Use getPrimaryAddress()->setFirstName() instead
+     *
      * @param string $value
      *
      * @return $this
      */
     public function setFirstName($value)
     {
-        return $this->setAttribute('firstName', $value);
+        $primaryAddress = $this->getPrimaryAddress();
+
+        if ($primaryAddress !== null) {
+            $primaryAddress->setFirstName($value);
+        } else {
+            $primaryAddress = Address::createFromData(['firstName' => $value]);
+        }
+
+        $this->setPrimaryAddress($primaryAddress);
+        $this->setAttribute('firstName', $value);
+
+        return $this;
     }
 
     /**
+     * @deprecated Use getPrimaryAddress()->getLastName() instead
+     *
      * @return string
      */
     public function getLastName()
     {
-        return $this->getAttribute('lastName');
+        return $this->getPrimaryAddress() !== null ? $this->getPrimaryAddress()->getLastName() : null;
     }
 
     /**
+     * @deprecated Use getPrimaryAddress()->setLastName() instead
+     *
      * @param string $value
      *
      * @return $this
      */
     public function setLastName($value)
     {
-        return $this->setAttribute('lastName', $value);
+        $primaryAddress = $this->getPrimaryAddress();
+
+        if ($primaryAddress !== null) {
+            $primaryAddress->setLastName($value);
+        } else {
+            $primaryAddress = Address::createFromData(['lastName' => $value]);
+        }
+
+        $this->setPrimaryAddress($primaryAddress);
+        $this->setAttribute('lastName', $value);
+
+        return $this;
     }
 
     /**
@@ -126,6 +195,29 @@ final class Customer extends Entity
      */
     public function setPrimaryAddress($value)
     {
+        // TODO: Remove after deprecated setEmail(), setFirstName(), setLastName() methods removed
+        if (is_array($value)) {
+            $value = Address::createFromData($value);
+        }
+
+        if (empty($value->getEmails()) && $this->getAttribute('email') !== null) {
+            $email = Email::createFromData([
+                'label' => 'main',
+                'value' => $this->getAttribute('email'),
+                'primary' => true,
+            ]);
+
+            $value->setEmails([$email]);
+        }
+
+        if (empty($value->getFirstName()) && $this->getAttribute('firstName') !== null) {
+            $value->setFirstName($this->getAttribute('firstName'));
+        }
+
+        if (empty($value->getLastName()) && $this->getAttribute('lastName') !== null) {
+            $value->setLastName($this->getAttribute('lastName'));
+        }
+
         return $this->setAttribute('primaryAddress', $value);
     }